move eventBus, appStorage, EditorStore to services directory
